# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Got Journey Thailand API",
    "description": "Travel Booking Platform API - Hotels, Cars, and Package Bookings",
    "version": "1.0.0",
    "contact": {
      "name": "Got Journey Thailand",
      "email": "support@gotjourneythailand.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development Server"
    },
    {
      "url": "https://api.gotjourneythailand.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/api/payments": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get all payments (Admin only)",
        "security": [{ "AdminAuth": [] }],
        "responses": {
          "200": { "description": "Successful response" }
        }
      }
    },
    "/api/payments/stats": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get payment statistics (Admin only)",
        "security": [{ "AdminAuth": [] }],
        "responses": {
          "200": {
            "description": "Statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stats": {
                      "type": "object",
                      "properties": {
                        "payments": { "type": "object" },
                        "bookings": { "type": "object" },
                        "revenue": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Get reviews",
        "parameters": [
          { "name": "item_type", "in": "query", "schema": { "type": "string", "enum": ["hotel", "car"] } },
          { "name": "item_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Successful response" }
        }
      },
      "post": {
        "tags": ["Reviews"],
        "summary": "Create a review (requires auth)",
        "security": [{ "UserAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReviewInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Review created" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_token"
      },
      "UserAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token"
      }
    },
    "schemas": {
...
      "ReviewInput": {
        "type": "object",
        "required": ["item_type", "item_id", "rating"],
        "properties": {
          "item_type": { "type": "string", "enum": ["hotel", "car"] },
          "item_id": { "type": "string" },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "comment_th": { "type": "string" },
          "comment_en": { "type": "string" }
        }
      },
...
    }
  }
}